// src/pages/profile/manage/type.tsx
export type ListingProfile = {
  id: string;
  fullName: string;
  specialization: string;
  profilePicture: string;
  overallRating: number;
  totalReviews: number;
  bio: string;
  baseLocation: string;
  serviceRadius: string;
  serviceAreas: string[];
  contactMethods: string[];
  businessLegalName: string;
  businessRegistration: string;
  yearsInBusiness: number;
  skills: string[];
};